# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API server"
        }
    ],
    "paths": {
        "/rating": {
            "post": {
                "tags": [
                    "Notation"
                ],
                "summary": "Noter un voyage",
                "operationId": "abdaadda14312676e1da4b96afd9f227",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trip_id",
                                    "rating"
                                ],
                                "properties": {
                                    "trip_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "description": "Note entre 1 et 5",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notation ajoutée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notation ajoutée avec succès."
                                        },
                                        "rating": {
                                            "$ref": "#/components/schemas/Rating"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/rating/{trip_id}": {
            "get": {
                "tags": [
                    "Notation"
                ],
                "summary": "Obtenir les notations pour un voyage",
                "operationId": "98208b1140f5a1939e4aac0be2442326",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "path",
                        "description": "ID du voyage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notations récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rating"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Voyage non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur interne"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Rating": {
                "required": [
                    "user_id",
                    "trip_id",
                    "rating"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "trip_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "description": "Note entre 1 et 5",
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}